lamp: linux wrappers for xpm and morse renderers
This uses lua-cqueues to provide an implementation of the lamp renderer
as a kind of shell pipeline tool.  I am sure things could be improved,
but this is a good start and lets me carry a virtual lamp on my desktop
as well as a slightly different piece of kit on my physical desk.