ws2812 vs gpio: down the AND gate "later"
Right now, the ws2812 driver doesn't wait for the bit stream to return
control to Lua, so rather than immediately kick the AND gate off, set up
a timer to do it "soon".  This is not really the right fix, since it
leaves the AND gate high for a long while, but it works well enough I
suppose.